CMake: Avoid creation of XXX_other_files targets
The XXX_other_files targets were introduced to collect files that are
added to resources and expose them to IDEs.  This resembles what we did
in qmake with OTHER_FILES but has the downside of cluttering the IDE
with additional targets.

The idiomatic way of exposing such files to IDEs is to add them as
source files to the related target and mark them as HEADER_FILE_ONLY.

We must avoid marking source files as HEADER_FILE_ONLY that are
additionally used in SOURCES.  This can only be done reliably when
finalizing a target.

Targets that are not finalized, for example targets that are added with
add_executable instead of qt_add_executable, are handled as before: for
those we still create the XXX_other_files target.  The same applies if
CMake < 3.19 is used.

Task-number: QTBUG-99808
Change-Id: I7220e91d726e9409d557d5cd33d171528b9d4a63
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>