Use the correct way to include qsimd_x86_p.h
qsimd_x86_p.h is not synced and installed as part of macOS multi-arch
builds, because of TEST_architecture_arch. With new approach when we
supposed to deploy only header files that belong to a platform in Qt
builds and installation, this file is missing. Considering
CMAKE_OSX_ARCHITECTURES in qt_internal_extend_target CONDITION fixes
this behavior and the file is added to CMake source tree.

Include qsimd_x86_p.h using private module include path, but not by
the relative path. This helps to detect the issue if file doesn't
exists in build or install directory when building qtbase, but
not other repositories.

The current architecture_arch test returns the very first
architecture that pass the check. Since arm64 is located in very
beginning it's returned first and other architectures that might be
used for multi-arch builds are not taken into account by conditions
that checks architecture_arch test results.

Fixes: QTBUG-106462
Task-number: QTBUG-103196
Change-Id: I6a9f110007d7069222c7d63bcd3ea586efab3aae
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>