AtSpiAdaptor: sync Cache signatures to at-spi2-core 2.45.91
Those signatures are the ones used by Gtk for some time now.  For Qt
it does not make a huge difference, as GetItem returns an empty list
and the signals are never emited.

The Cache.xml copy had to include a workaround: adding back the
direction on signal args, without which qdbusxml2cpp.cpp seems to
mistake those for direction="out" and requests "Out0" annotations.

Signed-off-by: Yann Dirson <ydirson@free.fr>