Clarify that FramelessWindowHint does not guarantee fixed sized windows
Depending on the window manager or operating system the act of making
a window borderless may also affect the ability for the user to move or
resize a window interactively. However, this is a side effect of making
a window borderless, and should not be relied on. The correct way to
ensure a window has a fixed size it to use setMinimum/MaximumSize.

Fixes: QTBUG-103503
Fixes: QTBUG-106710
Pick-to: 6.4
Change-Id: I6d82380f5a9af1f08b12cb445792ecbf71bb0377
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>